+++ src/main/java/com/Hechuyengia/Chuandoanbenh/service/DiagnosisService.java
@@ -51,7 +51,7 @@ public class DiagnosisService {
             String ma_benh = result[0].toString();
             String ten_benh = result[1].toString();
             //String loai_he = result[2].toString();
-           int ma_benh_int = Integer.parseInt(ma_benh);
+           Long ma_benh_int = Long.parseLong(ma_benh);
             BenhEntity benhEntity = new BenhEntity(ma_benh_int, ten_benh);
             benhEntities.add(benhEntity);
         }
@@ -66,7 +66,7 @@ public class DiagnosisService {
             String ma_benh = result[0].toString();
             String ten_benh = result[1].toString();
             //String loai_he = result[2].toString();
-            int ma_benh_int = Integer.parseInt(ma_benh);
+            Long ma_benh_int = Long.parseLong(ma_benh);
             BenhEntity benhEntity = new BenhEntity(ma_benh_int, ten_benh);
             benhEntities.add(benhEntity);
